add event listener. comments related to unknown issue
I am trying to declare the function showDetails that logs the name of the pokemons in the console with the event listener. However, when I do, the console shows "undefined" everytime I press the buttons. on the other hand, if I use an anonymous function within the eventlistener that logs in the console the pokemon name, it works absolutely fine.